Create GameOfPrimes
In the Primeland, there lives a monk who has a special power to check the integrity of a couple. The integrity of a couple can be expressed as a function of the primality of their love attribute. 

Couple (x,y) has the integrity as follows ( x and y are their respective attributes): 
If x and y are both prime, then it is x+y
If only x is prime, then it is x
If only y is prime, it is y
Else, it is x*y

Guideline :
Make sure that you code using FUNCTIONS to obtain maximum marks. YOU WILL LOSE MARKS IF THIS GUIDELINE IS NOT FOLLOWED

Input Format:
Two space separated positive integers, x and y.

Output Format:
A single line consisting of the integrity of couple. 

Constraints:
2≤x,y≤104

Example:
Input : 
5 9
Output : 
5

Input : 
4 100
Output : 
400