Create CreativeCrypts
Pragya has always been inspired by Julius Caesar's cryptic skills. In order to communicate with her friend, Naman, she decides to use a cyclic cipher. It is a type of substitution cipher in which each
letter in the text is replaced by a letter some fixed number of positions up or down the alphabet. Intimidating though it seems, Naman luckily has the blessings of your programming skills!

Help Naman by writing a C program to implement the cipher technique. Your program should ask the user for an input message, a shift parameter, and if it needs to be encrypted (E) or decrypted (D). Output should be the corresponding encoded or decoded message.

Note :
(1) Assume that a valid input message consists of only uppercase or lowercase alphabets. Uppercase alphabets are encoded/decoded using uppercase, and so on. 
(2) You may getchar() to input the message character by character, and putchar() to print the output.

E.g., With a shift parameter of 3, decrypting the input message “BPZ” should output “ESC” (B shifted right by 3 places is E, P is S, and Z is C).
Similarly, encrypting the message “ESC” should print the output “BPZ” (E shifted left by 3 places is B, and so on.)

Guideline: 
Make sure that you code using FUNCTIONS to obtain maximum marks. YOU WILL LOSE MARKS IF THIS GUIDELINE IS NOT FOLLOWED.

Input Format:
Input consists of 3 lines.
First line consists of a single integer N, the shift parameter.
Second line consists of a single character (E or D) telling you whether the message needs to be encrypted or decrypted respectively.
Third line consists of the input message, which is a sequence of characters.

Output Format:
A single line consisting of the output message.

Constraints:
0≤N≤103

Example:
INPUT:
2
D
FCJN
OUTPUT:
HELP

INPUT:
3
E
please
OUTPUT:
mibxpb