Create PerCapitaIncomeMdn
Vipul is a first-year student. He has always topped in his class, and wants to do the same here. But after coming to Kanpur, he sees a large section of population reeling under poverty. He realizes that he was disillusioned by growing per capita income in the country which is due to the widening gap between the rich and the poor. On further research, he finds out that “India's top 1% bag 73% of the country's wealth” and immediately comes to the conclusion that per capita income(mean) is not a good statistical measure in such data and we need to have a better measure for it like median or middle income. He needs your help in finding out if the median income from a sample sorted array of incomes given.

Median Calculation:
Median is calculated as follows. If the number of samples are odd, it’s simply the middle element in the sorted list. Else, it’s the average of the middle two elements in a sorted list.


INPUT:
The input consists of two lines. The first line mentions N(number of people in the sample)
The second line mentions N space-separated natural numbers which lists their respective annual incomes. 


OUTPUT:
Output the calculated median as a natural number. 

EXAMPLE:
Input -
5
10000 12000 15000 18000 6000000

Output - 
15000


Note: Few marks will be reserved for good programming practices.