Create OurExistence-CharArr
In an array of characters, verify if there exists an alphabet.
The alphabets may be in lower or upper case.


INPUT:
The first line of input contains the size n of the array. Then n lines of input follow giving the contents of the array.



OUTPUT:
You must output “Yes” or “No” if there exists an element with the desired property or not respectively in the given array.

CONSTRAINTS:
The size of the array is smaller than 20
The contents of the array are characters which may include lower case alphabets, upper case alphabets, numerals, visible special characters (ASCII range from 33 to 47 only)


EXAMPLE:
INPUT:
4
a
A
$
%
OUTPUT: 
Yes

Note: Few marks will be reserved for good programming practices.