Create LostInSpace
Will is playing in the Alpha Centauri with his new friend, the Alien robot. Both of them throw two clay balls in the same direction. Assume that these two balls, moving in the same direction collide inelastically (i.e., after the collision, the two balls stick to each other). You, being the genius programmer behind the Alien robot, would like to write a program to find the combined velocity of the new mass. You will be provided with the masses of the two clay balls and their velocities.

Vijay Dhar, gave you the mathematical equation to find this velocity, based on conservation of momentum: m1v1+m2v2=(m1+m2)v. In this equation m1,m2 are the masses of the two balls, v1,v2 are their velocities, and v is the final velocity of the combined mass.

NOTE: You can assume that the answer will always be an integer.

INPUT:
The first line of the input contains two space-separated integers: m1,m2, denoting masses of the two clay balls. The second line of the input contains two space-separated integers: v1,v2 denoting the velocities, respectively.

OUTPUT:
Single line representing the combined velocity(v).