Create SmartJudy
In Alpha Centauri, laws of physics are the same, but the universal constants are a bit different. Judy is feeling cold, so she is planning to make an electric heater. Before designing the heater, she wants to calculate the power consumption of this heater, so that she can rechange the solar batteries with the correct voltage. Unfortunately, she has to go on a date with Vijay, can you help her in calculating the power consumed by the heater coil?

Some formulas from Electrodynamics to help you out are:
1. P=I2R
2. R=ρlA
In these equations ρ, is the resistivity constant, ρ=100 in Alpha Centauri. R is the resistance of the coil. l, A are the length and the cross-sectional area of the coil. I is the current flowing through the coil, and P is the power consumption by the heater coil.

NOTE: You can assume that the answer will always be integral. All the values are given in standard units.

INPUT:
Single line containing three space-separated integers: I, L, A.

OUTPUT:
Single line denoting power consumption by the heater.

EXAMPLE:
INPUT:
2 3 50

OUTPUT:
24