Create StudentSearch
Do you know that student search internally uses a string of length 4(secret ID), to represent each student? This secret ID is extracted from the roll number of the student. You would like to be a part of the programming club and to impress the coordinators; you are planning to write a program, which outputs the secret ID, using the roll number.

In the hypothetical iitk, we have eight-digit roll numbers. Encoding to get the secret ID is defined as follows:
1. Let x denote the roll number of a student, and a, b, c, d be the numbers formed by taking two consecutive digits from x. For example, if x = 12131415 then a = 12, b = 13, c = 14, d = 15.
2. for each of the numbers a, b, c, d do the following
    a. Take modulo with 26
    b. get the character representation (if the number = 1, character representation = 'B', if the number = 25, character representation = 'Z')
3. These character representations of a, b, c, d represent the secret ID.

INPUT:
Single integer, denoting the roll number of a student.

OUTPUT:
Secret ID corresponding to that student.

EXAMPLE:
INPUT:
10111213

OUTPUT:
KLMN

Explanation:
In this example a = 10, b = 11, c = 12, d = 13. After taking modulo with 26, all of them remain the same. Their character representation gives "KLMN".