Create ThorVsThanos
Thor, the God of Thunder, to eliminate Thanos, created a lightning attack at a point on a planar battlefield. It is known that lightning affects only the area within a distance of p(power of Thor) from attack point.
You are given the attack point (x,y) and power(p) of Thor. Check if Thanos at position(a, b) will be damaged or not.

INPUT:
A single line contains 5 space-separated integers 
x y p a b

OUTPUT:
A single line with output as “AVENGERS” if Thanos is damaged, otherwise “THANOS”.

CONSTRAINTS:
-1000 ≤x,y,a,b≤ 1000
0≤r≤1000

EXAMPLE:
Input:
0 0 5 4 2
Output:
AVENGERS