Create ClientServer
“Lengthy problems are not always difficult”   -- Anonymous

Status codes are issued by a server in response to a client's request made to the server. The first digit of the status code specifies one of five standard classes of responses.

1xx : informational response – the request was received, continuing process

2xx : successful – the request was successfully received, understood and accepted

3xx : redirection – further action needs to be taken in order to complete the request

4xx : client error – the request contains bad syntax or cannot be fulfilled

5xx : server error – the server failed to fulfill an apparently valid request


Given the status code of some HTTP request, you need to tell which class the code belongs to. Moreover, being a network geek, you want to know the detailed type of client error if the error code belongs to the client error class. 

400 : "Bad Request"

401 : "Unauthorised"

402 : "Payment Required"

403 : "Forbidden"

404 : "Not found"

4xx and nothing of above : “Unknown”


Given an integer, print the class of the status code. Further, if the status code belongs to the client error class, then print the type of client error in the next line.
Note: Input status code is guaranteed to belong to one of the five standard classes mentioned above.

INPUT:
A single line consisting of a single integer, code.

OUTPUT:
First line containing "informational response", "successful", "redirection", "client error" or "server error" according to the class of response.
Second line containing the detailed type of client error if the code is of client error class. Else it is empty.

CONSTRAINT:
100 <=  code <= 599

EXAMPLE:
INPUT:
404
OUTPUT:
client error
Not found

INPUT:
201
OUTPUT:
successful