Create GameOfRoots
In a mystical world, Alice is lost somewhere in the woods. She reaches a point where she is able to see 2 trees. She wonders if they are real or imaginary. A creature appears to help her and tells her that the root of a tree actually represents the nature of the tree(real or imaginary). The creature also tells her that, the roots of equations ax2+bx+c=0 represent the actual roots of trees.

You have to check if a quadratic equation, ax2+bx+c=0, has :

 
 2 real and distinct roots

 
 2 real but same roots, or

 
 Both imaginary roots


INPUT:
A single line consisting of 3 space-separated floating-point numbers a, b and c.

OUTPUT:
Print a single line with “REAL AND DISTINCT”, “REAL BUT SAME” or “IMAGINARY” (without quotes) corresponding to cases (1), (2) and (3) respectively.

CONSTRAINTS:
−104≤a,b,c≤104

EXAMPLE:
INPUT:
1 -3 4
OUTPUT:
IMAGINARY

INPUT:
1 0 0
OUTPUT:
REAL BUT SAME