Create InvertMyWindow
Given a sequence of numbers a1,a2,..,an, a contiguous subsequence of length k is ai,ai+1,..,ai+k−1 such that i≥1 and i+k−1≤n.

Given an sequence of integers a1,a2,...,an, the number of inversions is defined as the number of pairs (i,j) such that i<j and ai>aj.

Consider the set of all k length contiguous subsequences of a given array. You have to find the maximum number of inversions possible in any such k length contiguous subsequence.

INPUT:

The first line contains the n - number of integers in the sequence. (n≤1000)
The second line contains the value of k as described above. (k≤n)
The next line contains n space separated integers of the sequence. 

OUTPUT:
Output the maximum number of inversions possible in any k length contiguous subsequence of the given sequence. (See the second question to know the meaning of inversions).

EXAMPLE:
INPUT: 
5
3
1 4  3 2  5
OUTPUT:
3
Explanation:
The subsequences are 1,4,3   4,3,2   3,2,5   which have 1,3 and 1 inversions respectively. 3 is the maximum amongst these which is the answer.
Note: 
Take the input in the global array as given in the template.
You have to implement the function getInversions as given in the template which will take l and r and give the number of inversions  in the sequence arr[l],arr[l+1],..,arr[l+r−1]. If you do not use this function to find the output, you get a zero.